football/requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
Babel==2.2.0
beautifulsoup4==4.4.1
billiard==3.3.0.22
boto==2.38.0
celery==3.1.19
coverage==4.0.3
decorator==4.0.6
Django==1.9.1
django-allauth==0.24.1
django-autoslug==1.9.3
django-braces==1.8.1
django-coverage-plugin==1.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-floppyforms==1.6.0
django-mailgun==0.8.0
django-model-utils==2.4
django-redis==4.3.0
django-secure==1.0.1
django-storages-redux==1.3
django-test-plus==1.0.11
docutils==0.12
et-xmlfile==1.0.1
factory-boy==2.6.0
fake-factory==0.5.3
gevent==1.0.2
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.4.1
httplib2==0.9.2
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
jdcal==1.2
Jinja2==2.8
kombu==3.0.33
MarkupSafe==0.23
nflgame
oauthlib==1.0.3
openpyxl==2.3.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.1.0
psycopg2==2.6.1
ptyprocess==0.5
Pygments==2.0.2
python-openid==2.2.5
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.0
requirements==0.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
sqlparse==0.1.18
traitlets==4.0.0
unicode-slugify==0.1.3
Unidecode==0.4.18
Werkzeug==0.11.2
whitenoise==2.0.6
